+++ packages/blade/src/components/ActionList/ActionListItem.tsx
@@ -23,6 +23,7 @@ import size from '~tokens/global/size';
 import type { DropdownProps } from '~components/Dropdown';
 import type { StringChildrenType, TestID } from '~src/_helpers/types';
 import { assignWithoutSideEffects } from '~src/utils/assignWithoutSideEffects';
+import { useTheme } from '~components/BladeProvider';
 
 type ActionListItemProps = {
   title: string;
@@ -303,9 +304,13 @@ const _ActionListItem = (props: ActionListItemProps): JSX.Element => {
     setShouldIgnoreBlur,
     setShouldIgnoreBlurAnimation,
     selectionType,
+    setIsOpen,
     dropdownTriggerer,
     isKeydownPressed,
   } = useDropdown();
+
+  const { platform } = useTheme();
+  const isMobile = platform === 'onMobile';
 
   const renderOnWebAs = props.href ? 'a' : 'button';
 
@@ -354,17 +359,25 @@ const _ActionListItem = (props: ActionListItemProps): JSX.Element => {
           disabled: props.isDisabled,
         })}
         {...makeActionListItemClickable((e: React.MouseEvent<HTMLButtonElement>): void => {
-          if (typeof props._index === 'number') {
-            onOptionClick(e, props._index);
+          if (props.onClick) {
+            props.onClick({ name: props.value, value: isSelected });
+            setIsOpen(false);
+          } else {
+            // Let my if be lonely its readable 😤
+            // eslint-disable-next-line no-lonely-if
+            if (typeof props._index === 'number') {
+              onOptionClick(e, props._index);
+            }
           }
-          props.onClick?.({ name: props.value, value: isSelected });
         })}
         {...metaAttribute({ name: MetaConstants.ActionListItem, testID: props.testID })}
         onMouseDown={() => {
-          setShouldIgnoreBlur(true);
-          // We want to keep focus on Dropdown's trigger while option is being clicked
-          // So We set this flag that ignores the blur animation to avoid the flicker between focus out + focus in
-          setShouldIgnoreBlurAnimation(true);
+          if (dropdownTriggerer === 'SelectInput') {
+            setShouldIgnoreBlur(true);
+            // We want to keep focus on Dropdown's trigger while option is being clicked
+            // So We set this flag that ignores the blur animation to avoid the flicker between focus out + focus in
+            setShouldIgnoreBlurAnimation(true);
+          }
         }}
         onMouseUp={() => {
           // (Contd from above comment...) We set this flag back to false since blur of SelectInput is done calling by this time
@@ -378,6 +391,7 @@ const _ActionListItem = (props: ActionListItemProps): JSX.Element => {
         intent={props.intent}
         isSelected={isSelected}
         isKeydownPressed={isKeydownPressed}
+        isMobile={isMobile}
       >
         <ActionListItemBody
           selectionType={selectionType}