+++ packages/razorsharp/src/code/blade/components/badge/index.ts
@@ -1,0 +1,1 @@
+export { transformBadge } from './badge';